add rest for module `USER` with test

+++ src/main/java/com/myChef/services/UserService.java
@@ -4,15 +4,17 @@
  */
 package com.myChef.services;
 
+import com.myChef.JPA.ChefDetails;
+import com.myChef.JPA.Ugroup;
+import com.myChef.JPA.User;
+import com.myChef.JPA.UserDetails;
+import com.myChef.repository.ChefDetailsRepository;
+import com.myChef.repository.UserDetailsRepository;
+import com.myChef.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ua.cn.al.teach.library2.jpa.Appuser;
-import ua.cn.al.teach.library2.jpa.Ugroup;
-import ua.cn.al.teach.library2.jpa.Userdetails;
-import ua.cn.al.teach.library2.repository.UserRepository;
-import ua.cn.al.teach.library2.repository.UserdetailsRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,44 +26,72 @@ import java.util.List;
 
 @Service
 public class UserService {
-private static final Logger logger =  LoggerFactory.getLogger(UserService.class);   
+private static final Logger logger =  LoggerFactory.getLogger(UserService.class);
 
-@Autowired
-UserRepository userRepository;
-@Autowired
-UserdetailsRepository detailsRepository;
-  
-  public List<Appuser> getAllUsers(){
-      return  userRepository.findAll();
-  }
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private UserDetailsRepository detailsRepository;
+    @Autowired
+    private ChefDetailsRepository chefDetailsRepository;
 
-    public Appuser getUserById(Long id) {
-        Appuser u = userRepository.findOne(id);
-        return u;
-    }
-    
-    public List<Appuser> findUserByName(String firstName, String lastName){
-        List<Userdetails> udl = detailsRepository.findByFirstNameAndLastName(firstName, lastName);
-        List<Appuser> res = new ArrayList<>();
-        udl.forEach((ud) -> {
-            res.add(userRepository.findOne(ud.getUserId()));
-        });
-        return res;
+    public List<User> getAllUsers() {
+        return userRepository.findAll();
     }
 
-    public Appuser addUser(Appuser au) {
-        logger.debug("Adding users %s with id %s", au.getUsername(), au.getUserId());
-        au = userRepository.save(au);
-        return au;
+
+    public List<User> getAllChefs() {
+        List<User> users = new ArrayList<>();
+
+        for (ChefDetails chefDetails : chefDetailsRepository.findAll()){
+            users.add(chefDetails.getUser());
+        }
+
+        return users;
     }
 
-    public void delUser(Long id){
-        Appuser u = userRepository.findOne(id);
-        if(u!=null){
-        logger.debug("Deleting users %s with id %s", u.getUsername(), u.getUserId());
-           List<Ugroup> gl = u.getUgroupList();
-           detailsRepository.delete(id);
-           userRepository.delete(id);
+    public List<User> getChefsByCity(long cityID) {
+        List<User> users = new ArrayList<>();
+
+        for (UserDetails userDetails : detailsRepository.findByCity_CityIdAndChefTrue(cityID)){
+            users.add(userDetails.getUser());
+        }
+
+        return users;
+    }
+
+    public User getUserById(Long id) {
+        User user = userRepository.findOne(id);
+        return user;
+    }
+
+    public User addUser(User user) {
+        logger.debug("Adding user %s with id %s", user.getUserName(), user.getUserId());
+        user = userRepository.save(user);
+        return user;
+    }
+
+    public User updateUser(User user) {
+        logger.debug("Updating user %s with id %s", user.getUserName(), user.getUserId());
+        user = userRepository.save(user);
+        return user;
+    }
+
+    public void delUser(Long id) {
+        User user = userRepository.findOne(id);
+        if (user != null) {
+            logger.debug("Deleting users %s with id %s", user.getUserName(), user.getUserId());
+
+//            List<Ugroup> gl = user.getUgroups();
+//            for (Ugroup ugroup : gl) {
+//                user.deleteGroup(ugroup);
+//            }
+
+            if (user.getChefDetails() != null) {
+                chefDetailsRepository.delete(id);
+            }
+            detailsRepository.delete(id);
+            userRepository.delete(id);
         }
     }
 }
